modification du path pour respect l'archi du projet + premire version de swagger modifié
