+++ Frontend_AgenceBank/src/pages/PrivateRoutes/AccountPage/AccountPage.module.css
@@ -1,0 +1,72 @@
+.main{
+  display:flex;
+  flex-direction:column;
+  justify-content:space-between;
+  align-items:center;
+  width:100%;
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  font-weight:bold;  
+  color: #2c3e50;
+  flex:1;
+  background-color:#12002b;
+}
+
+.available{
+  display:flex;
+  flex-direction:column;
+  justify-content:center;
+  align-items:center;
+  width:100%;
+  margin-top:0px;
+  border: 1px solid #2c3e50;
+  padding:25px 50px;
+  background-color:#fff;
+}
+
+.transactionSection{
+  display:flex;
+  flex-direction:column;
+  width:80%;
+  margin-bottom:20px;
+  padding-top:50px;
+  flex-grow:1;
+}
+
+.title{
+  display:flex;
+  justify-content:center;
+  align-items:center;
+  width:100%;
+  font-size:16px;
+  font-weight:bold;
+  margin-bottom:20px;
+}
+
+.amount{
+  display:flex;
+  justify-content:center;
+  align-items:center;
+  width:100%;
+  font-size:24px;
+  font-weight:bold;
+  margin-bottom:20px;
+  text-shadow: 0px 0px 5px #fff;
+}
+
+.header{
+  display:flex;
+  justify-content:space-between;
+  font-size:8px;
+  font-weight:bold;
+  color : #fff;
+  padding: 10px 0;
+  text-transform:uppercase;
+  
+}
+
+.header h2{
+  display:flex;
+  justify-content:center;
+  align-items:center;
+  width:25%;
+}
