+++ Frontend_AgenceBank/src/pages/PrivateRoutes/ProfilePage/ProfilePage.module.css
@@ -1,72 +1,94 @@
-.main{
+
+
+.bg_dark{
   display:flex;
   flex-direction:column;
-  justify-content:space-between;
   align-items:center;
-  width:100%;
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  font-weight:bold;  
-  color: #2c3e50;
+  background-color: #12002b;
   flex:1;
-  background-color:#12002b;
 }
 
-.available{
-  display:flex;
-  flex-direction:column;
-  justify-content:center;
-  align-items:center;
-  width:100%;
-  margin-top:0px;
-  border: 1px solid #2c3e50;
-  padding:25px 50px;
-  background-color:#fff;
-}
-
-.transactionSection{
-  display:flex;
-  flex-direction:column;
-  width:80%;
-  margin-bottom:20px;
-  padding-top:50px;
-  flex-grow:1;
+.hero{
+  color:white;
+  margin-bottom:2rem;
+  height:200px;
 }
 
 .title{
-  display:flex;
-  justify-content:center;
-  align-items:center;
-  width:100%;
-  font-size:16px;
-  font-weight:bold;
-  margin-bottom:20px;
+  display: flex;
+  flex-direction: column;
+    font-size: 2em;
+    margin:30px auto;
+    font-weight: bold;
 }
 
-.amount{
-  display:flex;
+.editBtnContainer{
   justify-content:center;
   align-items:center;
-  width:100%;
-  font-size:24px;
-  font-weight:bold;
-  margin-bottom:20px;
-  text-shadow: 0px 0px 5px #fff;
 }
 
-.header{
-  display:flex;
-  justify-content:space-between;
-  font-size:8px;
-  font-weight:bold;
-  color : #fff;
-  padding: 10px 0;
-  text-transform:uppercase;
+
+
+.button{
   
+  border-color: #00bc77;
+    background-color: #00bc77;
+    color: #fff;
+    font-weight: bold;
+    padding: 10px;
 }
 
-.header h2{
-  display:flex;
-  justify-content:center;
+.userTitleName{
+  flex-direction:column;
+}
+
+
+.userTitleForm{
+  flex-direction:column
+}
+
+.userTitleInputContainer{
   align-items:center;
-  width:25%;
+  justify-content:center;
+  margin-bottom:1rem;
+}
+
+.userTitleButtonContainer{
+  display:flex;
+  
+  align-items:center;
+  justify-content:center;
+  margin:0 1rem;
+  margin-bottom:1rem;
+}
+
+.userTitleInput{
+  width:150px;
+  padding:0.5rem ;
+  margin:0 1rem;
+  border:2px solid #aeaff0;
+  outline:none;
+}
+
+.userTitleInput::-webkit-input-placeholder{
+  color:#46296f;
+  font-weight:bold;
+}
+
+
+
+
+.userTitleButton{
+  display:flex;
+  align-items:center;
+  justify-content:center;
+  width:100px;
+  padding:0.5rem 2rem;
+  margin:0 1rem;
+  border-color: #00bc77;
+    background-color: #00bc77;
+    color: #fff;
+    font-weight: bold;
+    padding: 10px;
+    text-align:center;
 }
