+++ Frontend_AgenceBank/src/pages/PrivateRoutes/ProfilePage/Transactions/Transaction.jsx
@@ -1,7 +1,7 @@
 import style from './Transaction.module.css'
 import { Link } from 'react-router-dom'
 
-export default function CountComponent({count, title, balanceType}) {
+export default function CountComponent({id ,count, title, balanceType}) {
   return (
     <section className={style.account}>
         <div className={style.accountContainer}>
@@ -10,7 +10,7 @@ export default function CountComponent({count, title, balanceType}) {
           <p className={style.accountDescription}>{balanceType}</p>
         </div>
         <div className={style.accountCta}>
-          <Link className={style.transactionBtn} to='/profile'>View transactions
+          <Link className={style.transactionBtn} to={`/account/${id}`} >View transactions
           </Link>
         </div>
       </section>
