--- Frontend_AgenceBank/src/pages/PrivateRoutes/UserPage/UserPage.module.css
@@ -1,94 +1,0 @@
-
-
-.bg_dark{
-  display:flex;
-  flex-direction:column;
-  align-items:center;
-  background-color: #12002b;
-  flex:1;
-}
-
-.hero{
-  color:white;
-  margin-bottom:2rem;
-  height:200px;
-}
-
-.title{
-  display: flex;
-  flex-direction: column;
-    font-size: 2em;
-    margin:30px auto;
-    font-weight: bold;
-}
-
-.editBtnContainer{
-  justify-content:center;
-  align-items:center;
-}
-
-
-
-.button{
-  
-  border-color: #00bc77;
-    background-color: #00bc77;
-    color: #fff;
-    font-weight: bold;
-    padding: 10px;
-}
-
-.userTitleName{
-  flex-direction:column;
-}
-
-
-.userTitleForm{
-  flex-direction:column
-}
-
-.userTitleInputContainer{
-  align-items:center;
-  justify-content:center;
-  margin-bottom:1rem;
-}
-
-.userTitleButtonContainer{
-  display:flex;
-  
-  align-items:center;
-  justify-content:center;
-  margin:0 1rem;
-  margin-bottom:1rem;
-}
-
-.userTitleInput{
-  width:150px;
-  padding:0.5rem ;
-  margin:0 1rem;
-  border:2px solid #aeaff0;
-  outline:none;
-}
-
-.userTitleInput::-webkit-input-placeholder{
-  color:#46296f;
-  font-weight:bold;
-}
-
-
-
-
-.userTitleButton{
-  display:flex;
-  align-items:center;
-  justify-content:center;
-  width:100px;
-  padding:0.5rem 2rem;
-  margin:0 1rem;
-  border-color: #00bc77;
-    background-color: #00bc77;
-    color: #fff;
-    font-weight: bold;
-    padding: 10px;
-    text-align:center;
-}
